orchestration/requirements.txt
prefect
# The orchestrator will need to call other scripts, so we need to ensure
# the environment can find them. The sys.path hack will be used.
python-logstash-async

# For gRPC
grpcio
grpcio-tools
